pyproject.toml
[project]
name = "govstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.1",
    "anthropic>=0.72.0",
    "asyncio-mqtt>=0.16.2",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.14.2",
    "boto3>=1.40.62",
    "chardet>=5.2.0",
    "chromadb>=1.3.0",
    "cohere>=5.20.0",
    "en-core-web-lg",
    "execnet>=2.1.1",
    "faker>=37.12.0",
    "fasta2a>=0.6.0",
    "fastapi>=0.120.2",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "grafana-api>=1.0.3",
    "groq>=0.33.0",
    "gunicorn>=23.0.0",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "langchain>=1.0.3",
    "llama-cloud>=0.1.35",
    "llama-index>=0.9",
    "llama-index-agent-openai>=0.4.12",
    "llama-index-core>=0.12.52.post1",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-groq>=0.3.2",
    "llama-index-llms-openai>=0.4.7",
    "llama-index-readers-file>=0.4.11",
    "llama-index-readers-llama-parse>=0.4.0",
    "llama-index-vector-stores-chroma>=0.4.2",
    "llama-parse>=0.6.54",
    "locust>=2.41.6",
    "logfire[aiohttp-client,all,fastapi,httpx,requests,system-metrics]>=4.14.2",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "markdownify>=1.2.0",
    "matplotlib>=3.10.7",
    "memory-profiler>=0.61.0",
    "minio>=7.2.18",
    "mistralai>=1.9.11",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "openai>=1.109.1",
    "opentelemetry-instrumentation-asyncpg>=0.59b0",
    "opentelemetry-instrumentation-dbapi>=0.59b0",
    "opentelemetry-instrumentation-llamaindex>=0.47.5",
    "opentelemetry-instrumentation-psycopg>=0.59b0",
    "paho-mqtt>=2.1.0",
    "pandas>=2.2.3",
    "presidio-analyzer>=2.2.360",
    "presidio-anonymizer>=2.2.360",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pydantic-ai>=1.9.0",
    "pydantic-settings>=2.11.0",
    "pympler>=1.1",
    "pypdf>=5.9.0",
    "pysqlite3-binary>=0.5.4",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.1.1",
    "pytest-metadata>=3.1.1",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "sentry-sdk>=2.43.0",
    "spacy>=3.8.7",
    "sqlalchemy>=2.0.44",
    "starlette>=0.49.1",
    "tabulate>=0.9.0",
    "tiktoken>=0.12.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "unstructured>=0.18.15",
    "uvicorn>=0.38.0",
    "vadersentiment>=3.3.2",
    "zope-event>=6.0",
    "zope-interface>=8.0.1",
]

[tool.uv.sources]
en-core-web-lg = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl" }
